[fix] Adjust project structure for admin Service and Repository
