+++ src/main/java/com/twm/repository/admin/PersonalityRepository.java
@@ -1,0 +1,12 @@
+package com.twm.repository.admin;
+
+import com.twm.dto.PersonalityDto;
+
+import java.util.List;
+
+public interface PersonalityRepository {
+
+    boolean savePersonality(PersonalityDto personalityDto);
+    List<PersonalityDto> getPersonality(Integer id);
+
+}
