+++ src/main/java/com/twm/service/chat/ChatService.java
@@ -13,10 +13,10 @@ public interface ChatService {
     List<ReturnCategoryDto> getAllCategoryButtons();
     String getUrlByCategory(Long categoryId);
     Map<String, Object> chat(Long userId, String sessionId, String question);
-    void saveButton(CreateButtonDto createButtonDto);
-    Map<String, Object> getButton(Integer id);
-    CreateButtonDto updateButton(CreateButtonDto createButtonDto);
-    boolean deleteButton(Long id);
-    Map<String, Object> savePersonality(PersonalityDto personalityDto);
-    Map<String, Object> getPersonality(Integer id);
+//    void saveButton(CreateButtonDto createButtonDto);
+//    Map<String, Object> getButton(Integer id);
+//    CreateButtonDto updateButton(CreateButtonDto createButtonDto);
+//    boolean deleteButton(Long id);
+//    Map<String, Object> savePersonality(PersonalityDto personalityDto);
+//    Map<String, Object> getPersonality(Integer id);
 }
