+++ src/main/java/com/twm/service/chat/impl/ChatServiceImpl.java
@@ -3,6 +3,7 @@ package com.twm.service.chat.impl;
 import com.twm.dto.*;
 import com.twm.dto.ReturnQuestionDto;
 import com.twm.dto.TypesDto;
+import com.twm.repository.admin.PersonalityRepository;
 import com.twm.repository.chat.ChatRepository;
 import com.twm.service.chat.ChatService;
 import com.twm.util.JwtUtil;
@@ -27,6 +28,8 @@ import java.util.*;
 public class ChatServiceImpl implements ChatService {
 
     private final ChatRepository chatRepository;
+
+    private final PersonalityRepository personalityRepository;
 
     @Resource
     private OpenAiChatModel openAiChatModel;
@@ -62,8 +65,8 @@ public class ChatServiceImpl implements ChatService {
         messages.add(new SystemMessage("這些是你們的對話紀錄 : " + sessionHistory));
 
         try {
-            log.info("人設: " + chatRepository.getPersonality(0));
-            messages.add(new SystemMessage("這是你的人設 : " + chatRepository.getPersonality(0)));
+            log.info("人設: " + personalityRepository.getPersonality(0));
+            messages.add(new SystemMessage("這是你的人設 : " + personalityRepository.getPersonality(0)));
             messages.add(new SystemMessage("常見問答都在此 : " + chatRepository.getFAQ()));
         }catch (RuntimeException e) {
             throw new RuntimeException("Failed to load agent", e);
@@ -109,52 +112,52 @@ public class ChatServiceImpl implements ChatService {
         return chatRepository.findAnswerByQuestion(buttonId);
     }
 
-    @Override
-    public void saveButton(CreateButtonDto createButtonDto){
-
-        chatRepository.saveButton(createButtonDto);
-
+//    @Override
+//    public void saveButton(CreateButtonDto createButtonDto){
+//
+//        chatRepository.saveButton(createButtonDto);
+//
+////        Map<String, Object> result = new HashMap<>();
+////        result.put("data", chatRepository.saveButton(createButtonDto));
+//
+////        return result;
+//    }
+//
+//    @Override
+//    public Map<String, Object> getButton(Integer id){
+//
 //        Map<String, Object> result = new HashMap<>();
-//        result.put("data", chatRepository.saveButton(createButtonDto));
-
+//        result.put("data", chatRepository.getButton(id));
+//
 //        return result;
-    }
+//
+//    }
+//
+//    @Override
+//    public CreateButtonDto updateButton(CreateButtonDto createButtonDto) {
+//        return chatRepository.updateButton(createButtonDto);
+//    }
+//
+//    @Override
+//    public boolean deleteButton(Long id) {
+//        return chatRepository.deleteButton(id);
+//    }
 
-    @Override
-    public Map<String, Object> getButton(Integer id){
-
-        Map<String, Object> result = new HashMap<>();
-        result.put("data", chatRepository.getButton(id));
-
-        return result;
-
-    }
-
-    @Override
-    public CreateButtonDto updateButton(CreateButtonDto createButtonDto) {
-        return chatRepository.updateButton(createButtonDto);
-    }
-
-    @Override
-    public boolean deleteButton(Long id) {
-        return chatRepository.deleteButton(id);
-    }
-
-    @Override
-    public Map<String, Object> savePersonality(PersonalityDto personalityDto) {
-        Map<String, Object> result = new HashMap<>();
-        result.put("result", chatRepository.savePersonality(personalityDto));
-
-        return result;
-    }
-
-    @Override
-    public Map<String, Object> getPersonality(Integer id) {
-        Map<String, Object> result = new HashMap<>();
-        result.put("data", chatRepository.getPersonality(id));
-
-        return result;
-    }
+//    @Override
+//    public Map<String, Object> savePersonality(PersonalityDto personalityDto) {
+//        Map<String, Object> result = new HashMap<>();
+//        result.put("result", chatRepository.savePersonality(personalityDto));
+//
+//        return result;
+//    }
+//
+//    @Override
+//    public Map<String, Object> getPersonality(Integer id) {
+//        Map<String, Object> result = new HashMap<>();
+//        result.put("data", chatRepository.getPersonality(id));
+//
+//        return result;
+//    }
 
     public List<ReturnCategoryDto> getAllCategoryButtons() {return chatRepository.findAllCategoryButtons();};
